Simulate FM pre-emphasis
I measured the demodulated audio from a TH-D7A
and found that the SPACE(2200Hz) tone was about
half the amplitude of the MARK(1200Hz) tone. I
created a new wavetable at half the amplitude
to use for MARK. No filtering or amplitude
matching is used, as the ST7 does not have
time for that. After doing so, and giving
the CC1050 more time to start up, I was
able to decode a packet with my TH-D7A.